Create access proxy server
This server provides the redirect URI that OAuth requires to hand over
the access code to request the API tokens so a Spotify connection can be
established. The access code endpoint is protected by an access key that
can be set on startup of the server to prevent unauthorized token access
from other sources able to call the access server endpoint.
Access Code Retrieval is handled via a HTTP calls between the Python
HTTP Server and the Javascript running in the quick access panel.
Port 49983 has been chosen as this seems to be an uncommonly used port
and will keep port collisions to a minimum. The host is set to bind to
localhost to prevent anyone from the network from being able to reach it
while it is up.
A ThreadedServer approach is used to be able to stop and start the
server on demand without blocking the Python's execution.