Create way to interface with Spotify
This is the beginning of a way of interfacing with Spotify. It includes
both a custom Spotify API Wrapper (to limit functionality to only the
necessary components relevant to Spotideck, as well as handle token
refreshing automatically) and a way to interface with the Access Server.

The Spotify Wrapper is incomplete but serves as a functioning proof of
concept and can easily be extended to include other required api calls.

The Access Server functions allow interfacing with the Python part of
the Plugin and give control over Access Server lifecycle, OAuth access
code retrieval and waiting for the Access Server to come online.
Once the access code and PCKE have been retrieved from Access Server the
`getTokens()` function allows retrieval of final set of tokens from the
Spotify API and allows instantiating a Spotify Client.

`setupSpotifyClient()` automatically handles Access Server lifecycle,
access code retrieval, token retrieval and Spotify instantiation.