Finish writing the functioning OAuth backend
Turns out that ThreadingHTTPServer creates a new instance of the
RequestHandler on each call. Who knew? This leads to persistance on the
self object being lost. Instead, a dict is used to store access codes
based on the access_protection_codes.

Multiple parts have also been changed to localize and instance variables
and make sure things are properly encapsulated.